+++ mmm_vadd_dfx/src/exp2_helper.cpp
@@ -2,7 +2,7 @@
 #include <sys/time.h>
 
 // Allocate memory on device and map pointers into the host
-void exp2_allocate_mem (cl_object &cl_obj, krnl_object &krnl_obj, float **ptr_out, int size_in_bytes) {
+void exp2_allocate_mem (cl_object &cl_obj, krnl_object &krnl_obj, int **ptr_out, int size_in_bytes) {
     cl_int err;
 
     // These commands will allocate memory on the Device. The cl::Buffer objects can
@@ -12,11 +12,11 @@ void exp2_allocate_mem (cl_object &cl_obj, krnl_object &krnl_obj, float **ptr_ou
     cl::Buffer *buf_out = &krnl_obj.buffers[0];
 
     //We then need to map our OpenCL buffers to get the pointers
-    OCL_CHECK(err, (*ptr_out) = (float*)cl_obj.q.enqueueMapBuffer (*buf_out , CL_TRUE , CL_MAP_READ , 0, size_in_bytes, NULL, NULL, &err)); // pointer to result out, should be read???
+    OCL_CHECK(err, (*ptr_out) = (int*)cl_obj.q.enqueueMapBuffer (*buf_out , CL_TRUE , CL_MAP_READ , 0, size_in_bytes, NULL, NULL, &err)); // pointer to result out, should be read???
 }
 
 // Unmap device memory when done
-void exp2_deallocate_mem (cl_object &cl_obj, krnl_object &krnl_obj, float *ptr_out) {
+void exp2_deallocate_mem (cl_object &cl_obj, krnl_object &krnl_obj, int *ptr_out) {
     cl_int err;
 
     cl::Buffer *buf_out = &krnl_obj.buffers[0];
